+++ ScreenSound.API/Requests/ArtistaRequest.cs
@@ -1,0 +1,5 @@
+﻿namespace ScreenSound.API.Requests
+{
+    public record ArtistaRequest(string nome, string bio);
+    
+}
