+++ ScreenSound.Shared.Modelos/Modelos/Genero.cs
@@ -1,0 +1,26 @@
+﻿using ScreenSound.Modelos;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ScreenSound.Shared.Modelos.Modelos
+{
+    // código omitido
+
+    public class Genero
+    {
+        public int Id { get; set; }
+        public string? Nome { get; set; } = string.Empty;
+        public string? Descricao { get; set; } = string.Empty;
+        public virtual ICollection<Musica> Musicas { get; set; }
+
+
+        public override string ToString()
+        {
+            return $"Nome: {Nome} - Descrição: {Descricao}";
+        }
+    }
+
+}
