+++ ScreenSound.Shared.Modelos/Modelos/Musica.cs
@@ -1,4 +1,5 @@
-﻿using System.Text.Json.Serialization;
+﻿using ScreenSound.Shared.Modelos.Modelos;
+using System.Text.Json.Serialization;
 
 namespace ScreenSound.Modelos;
 
@@ -15,6 +16,9 @@ public class Musica
     public int Id { get; set; }
     public int? AnoLancamento { get; set; }
     public virtual Artista? Artista { get; set; }
+
+    public virtual ICollection<Genero> Generos { get; set; }
+
 
     public void ExibirFichaTecnica()
     {
